ENhetstest, ny klass och kodstädning
