+++ PersonnummerValideringTest/packages.config
@@ -1,5 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Microsoft.CodeCoverage" version="17.12.0" targetFramework="net472" />
+  <package id="Microsoft.NET.Test.Sdk" version="17.12.0" targetFramework="net472" />
   <package id="MSTest.TestAdapter" version="2.2.10" targetFramework="net472" />
   <package id="MSTest.TestFramework" version="2.2.10" targetFramework="net472" />
 </packages>
